==> 00_Example_Processing.Rmd <==
---
title: "Stat_Project"
author: "Deniz KB"
date: '2022-05-07'
output: html_document
---

# Preprocessing

* Droping NAs, \
* Removing units of "max_power", "engine", "mileage", \
* Converting "max_power", "engine", "mileage" into numeric column, \
* Splitting "name" into "Make" and "Model" \
* Removing "Make" counts less than 10 \
* Removing "torque" and "Model" \
* "year"
  - for EDA: We keep year as it is.
  - for Regression: We convert "year" into "age" and remove "year".




```{r}
library(readr)
library(dplyr)
library(stringr)
library(tidyr)

car_details <- read_csv("car_details_v3.csv")
#Droping NAs
vehicles <- car_details %>% drop_na()

#run once
#Removing of the units of max_power, engine, mileage
vehicles$max_power <- substr(vehicles$max_power, 1, nchar(vehicles$max_power)-4)
vehicles$engine <- substr(vehicles$engine, 1, nchar(vehicles$engine)-3)
vehicles$mileage <- substr(vehicles$mileage, 1, nchar(vehicles$mileage)-5)

#Converting max_power, engine, mileage into numeric column
vehicles <- transform(vehicles, mileage = as.numeric(mileage))
vehicles <- transform(vehicles, engine = as.numeric(engine))
vehicles <- transform(vehicles, max_power = as.numeric(max_power))

#Splitting "name" into "Make" and "Model"
vehicles <- vehicles %>% mutate(name = str_replace(name, "\\s", "|")) %>% 
  separate(name, into = c("Make", "Model"), sep = "\\|")

#Checking for categories of "Make"
#table(vehicles$Make)
#Removing "Make" counts less than 10
vehicles <- subset(vehicles, !(Make %in% c('Ambassador', 'Ashok', 'Daewoo', 'Force', 'Isuzu', 'Kia', 'Land', 'MG', 'Opel')))

#Removing torque, Model
vehicles <- select(vehicles, -torque)
vehicles <- select(vehicles, -Model)

#We keep year for EDA; we convert it into age and remove year for regression.
vehicles$age <- 2022 - vehicles$year
vehicles <- select(vehicles, -year)

head(vehicles)
str(vehicles)
```

# Further Discovery

```{r}
#levels of Make
unique(vehicles$Make)

#levels of fuel
unique(vehicles$fuel)

#levels of seller_type
unique(vehicles$seller_type)

#levels of transmission
unique(vehicles$transmission)

#levels of owner 
unique(vehicles$owner)

#unique values of seats
sort(unique(vehicles$seats))

#unique values of age
sort(unique(vehicles$age))
```


```{r}
#Distribution of selling_price
hist(vehicles$selling_price)

#Distribution of km_driven
hist(vehicles$km_driven)

#Distribution of mileage
hist(vehicles$mileage)

#Distribution of engine
hist(vehicles$engine)

#Distribution of max_power
hist(vehicles$max_power)
```


# Fitting the initial Full Multiple Linear Regression Model

```{r}
#Multiple linear regression
mlr_obj <- lm(selling_price ~ .,
             data = vehicles)
summary(mlr_obj)
#full model is significant but there are some individual predictors are not significant.
```



# Checking for Collinearity 

"Make" has the greatest VIF score that is greater than 5. We remove this column.

```{r}
library(car)
vif(mlr_obj)  # Will throw an error due to perfect collinearity.
```


```{r}
reduced_1 <- lm(selling_price ~ .-Make,
             data = vehicles)
#summary(reduced_1)
```

After removing "Make", we still have a predictor that has VIF score greater than 5. We drop "engine" in next step.

```{r}
vif(reduced_1)
```

```{r}
reduced_2 <- lm(selling_price ~ .-Make-engine,
             data = vehicles)
#summary(reduced_2)
```

All predictors have VIF scores less than 5 in reduced model.

```{r}
vif(reduced_2)
```


# Variable Selection via Backward AIC

After handling collinearity, we will select the most efficient predictors for our model by using AIC. 

As we can see the outcomes of step() function, we dont need to drop any other predictors since AIC wont improve more.

```{r}
step(reduced_2)
```


# Significance of Categorical Variables

I just rename our final model and check for the model significance and individual predictor significance.

```{r}
# rename reduced_2 as final_model
final_model <- lm(selling_price ~ .-Make-engine,
             data = vehicles)
summary(final_model)
```

Our model is significant based on p value of F-statistics < 0.05 but we have some predictors are not significant. 

Now, we will test the significance of the individual predictors.

**Final Model vs Model without "fuel"**

"fuel" predictor is significant.

```{r}
#without fuel
null_model1 <- lm(selling_price ~ .-Make-engine-fuel,
             data = vehicles)

#full vs without fuel
anova(final_model, null_model1)
```


**Final Model vs Model without "owner"**

"owner" predictor is significant.

```{r}
#without owner
null_model2 <- lm(selling_price ~ .-Make-engine-owner,
             data = vehicles)

#full vs without owner
anova(final_model, null_model2)
```


**Final Model vs Model without "owner+ fuel"**

"owner + fuel" predictor is significant.

```{r}
#without owner + fuel
null_model3 <- lm(selling_price ~ .-Make-engine-owner-fuel,
             data = vehicles)

#full vs without fuel + owner
anova(final_model, null_model3)
```


# Final Multiple Linear Regression

```{r}
summary(final_model)
```

# Interpretations of Full Model

#Not sure if we are interpreting the final model of Expectation 3 or final model of Expectation 4???

__RSE= 454900__

On average, out predicted selling price miss the true selling price by 454,900 ???(what is the unit for selling price).


__R-squared= 0.6842__

Our model explainss 68.42% of variability in sellinng price.



# Expectations 4
# Non-linearity check -- Mileva's part

```{r}

## Residuals-vs-Fitted Plot (the 1st diagnostic plot):
plot(final_model, which = 1)
```




```{r}


```



```{r}



```


```{r}



```


```{r}



```
