==> 01_Processing.Rmd <==
---
title: "01 Processing Pipeline"
output: html_document
---

# Load Packages
```{r}
library(readr)
library(dplyr)
library(stringr)
library(tidyr)
library(tidyverse)
```

# Read Data
```{r}
car_details <- read_csv("data/car_details_raw.csv")
head(car_details)
```

# Check for NA Values
```{r}
# check for NAs in each column
colSums(is.na(car_details)) 
```

# Preprocessing

* Dropping NAs, \
* Removing units of "max_power", "engine", "mileage", \
* Converting "max_power", "engine", "mileage" into numeric column, \
* Converting selling price to dollars, \
* Splitting "name" into "Make" and "Model" \
* Grouping "Make" counts less than 10 into "Other" \
* Converting categorical variables in factor column \
* Removing "torque" and "Model" \
* "year"
  - for EDA: We keep year as it is.
  - for Regression: We convert "year" into "age" and remove "year".

```{r}
#Dropping NAs
vehicles <- car_details %>% drop_na()
```

```{r}
#Removing of the units of max_power, engine, mileage
vehicles$max_power <- substr(vehicles$max_power, 1, nchar(vehicles$max_power)-4)
vehicles$engine <- substr(vehicles$engine, 1, nchar(vehicles$engine)-3)
vehicles$mileage <- substr(vehicles$mileage, 1, nchar(vehicles$mileage)-5)

#Converting max_power, engine, mileage into numeric column
vehicles <- transform(vehicles, mileage = as.numeric(mileage))
vehicles <- transform(vehicles, engine = as.numeric(engine))
vehicles <- transform(vehicles, max_power = as.numeric(max_power))

head(vehicles)
```

```{r}
# Convert selling price to dollars
vehicles$selling_price_dollars = 0.013*vehicles$selling_price
head(vehicles)
```

```{r}
#Splitting "name" into "Make" and "Model"
vehicles <- vehicles %>% mutate(name = str_replace(name, "\\s", "|")) %>% 
  separate(name, into = c("Make", "Model"), sep = "\\|")

#Checking for categories of "Make"
table(vehicles$Make)
```

```{r}
#Collapsing "Make" counts less than 10 into "Other"
vehicles$Make <- fct_collapse(vehicles$Make, Other = c('Ambassador', 'Ashok', 'Daewoo', 'Force', 'Isuzu', 'Kia', 'Land', 'MG', 'Opel'))

#Checking for categories of "Make"
table(vehicles$Make)
```

```{r}
# Convert categorical variables into factors
vehicles$Make <- as.factor(vehicles$Make)
vehicles$fuel <- as.factor(vehicles$fuel)
vehicles$seller_type <- as.factor(vehicles$seller_type)
vehicles$transmission <- as.factor(vehicles$transmission)
vehicles$owner <- as.factor(vehicles$owner)

head(vehicles)
```

```{r}
#Removing torque, Model
vehicles <- select(vehicles, -torque)
vehicles <- select(vehicles, -Model)

head(vehicles)
```

```{r}
# Convert year to age
vehicles$age <- 2022 - vehicles$year

head(vehicles)
```

# Transform response selling_price into binary

```{r}
#Transformation:
#values over $10000: (No)
#values below $10000: (Yes)

THRESHOLD = 10000 

vehicles$selling_price_binary <- (vehicles["selling_price_dollars"] <= THRESHOLD)

vehicles["selling_price_binary"][vehicles["selling_price_binary"] == TRUE]  <- "Yes"
vehicles["selling_price_binary"][vehicles["selling_price_binary"] == FALSE]  <- "No"

vehicles$selling_price_binary <- as.factor(vehicles$selling_price_binary)

str(vehicles)
```

# Save processed data as csv
```{r}
write.csv(vehicles,"data/cars_processed.csv", row.names = FALSE)
```
