==> 03_Categorical_Variables.Rmd <==
---
title: "03 Categorical Variable Analysis"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Load Packages
```{r}
library(readr)
library(dplyr)
library(stringr)
library(tidyr)
library(tidyverse)
```

# Read Data
```{r}
# Read data
vehicles <- read_csv("data/cars_processed.csv")

# Convert chr columns to factors 
factor_column_names <- colnames(vehicles %>% select(where(is.character)))
vehicles[, factor_column_names] <- lapply(vehicles[,factor_column_names], factor)

head(vehicles)
```

# Contingency table
```{r}
price_transmission_table <- with(vehicles, table(transmission, selling_price_binary))
price_transmission_table
```

# Proportion table. 
The conditional proportion table indicates they are likely dependent. 
```{r}
with(vehicles, prop.table(price_transmission_table,1))

```

# Chi.sq test
```{r}
chisq.test(price_transmission_table, correct = FALSE)
pchisq(1564.5, df=1, lower.tail=F)
```

Chi-squared test statistic: 
$$\chi^2 = \sum_{ij}{\frac{(E_{ij} - Observed_{ij})^2}{E_{ij}}}$$

# Expected counts
```{r}
expected_counts <- function(X) { # X is a contingency table of arbitrary size
  
  row_sums <- matrix(rowSums(X), nrow = length(rowSums(X)))  # nrow x 1 matrix
  col_sums <- matrix(colSums(X), ncol = length(colSums(X))) # 1 x ncol matrix
  
  # Matrix of expected counts
  expected <- (row_sums %*% col_sums) / sum(X) 
  print(expected)
}
```

```{r}
expected_counts(price_transmission_table)
```
