==> 05_Logistic_Regression.Rmd <==
---
title: "Logistic_Regression"
author: "Marina Sanchez"
date: "5/13/2022"
output: html_document
---
```{r setup, include=FALSE} 
knitr::opts_chunk$set(warning = FALSE, message = FALSE) 
```

```{r}
library(readr)
library(dplyr)
library(stringr)
library(tidyr)

cars <- read_csv("data/cars_processed.csv")
head(cars)

cars$Make <- as.factor(cars$Make)
cars$fuel <- as.factor(cars$fuel)
cars$seller_type <- as.factor(cars$seller_type)
cars$transmission <- as.factor(cars$transmission)
cars$owner <- as.factor(cars$owner)
cars$selling_price_binary <- as.factor(cars$selling_price_binary)

summary(cars)
```

```{r}
hist(cars$selling_price_dollars)
```

```{r}
#Remove rows which selling_price_dollars is over 40000
#cars <- cars[!(cars$selling_price_dollars>40000),]
#hist(cars$selling_price_dollars)
```


```{r}
#Remove redundant response (different formats): selling_price_dollars, selling_price
#Only the categorical response is required: selling_price_dollars
#Remove also year, since we already have age

cars <- cars %>% select(-c(selling_price_dollars, selling_price, year))
summary(cars)
```

```{r}
library(ggplot2)
# Most basic bar chart
ggplot(cars, aes(x = selling_price_binary)) +
      geom_bar(fill = "coral") +
      theme_classic() +
    labs(
    title = "Affordability of the cars with a budget of $10,000",
  )

```


# Fit an initial multiple logistic regression with all of the predictors 
```{r}
glm.multiple <- glm(selling_price_binary ~ .,
                    data=cars,
                    family="binomial")

summary(glm.multiple)
```

# Deal with outliers
```{r, results = 'hide'}
## Influence via Cook's distance:
cooks.distance(glm.multiple)
plot(glm.multiple, which=4)
```

```{r}
#Removing observations 142, 2625 and 6896
cars <- cars[-c(144, 2635, 6923), ]

#Fit Logistic regression again
glm.multiple <- glm(selling_price_binary ~ .,
                    data=cars,
                    family="binomial")

summary(glm.multiple)
```

There are some predictors that look to be significant: MakeChevrolet, MakeTata, km_driven, seller_typeTrustmark Dealer, transmissionManual, ownerSecond Owner, mileage, max_power, seats and age   

# Using VIF to remove colinearity
```{r}
library(car)

vif(glm.multiple)
```
# We are removing Make since it has the highest VIF
```{r}
cars <- cars %>% select(-Make)

glm.multiple_red <- glm(selling_price_binary ~ .,
                    data=cars,
                    family="binomial")

vif(glm.multiple_red)
```
All the VIF values are below 5 now.

# Apply AIC for variable selection
```{r}
step(glm.multiple_red)
```
Our final model will include the predictors: km_driven + fuel + seller_type + transmission + owner + engine + max_power + age


#Fit logistic regression for our final model
```{r}
cars <- cars %>% select(-c(mileage, seats))

glm.multiple_final <- glm(selling_price_binary ~ .,
                      data=cars,
                     family="binomial")

summary(glm.multiple_final)
```
Most statistical significant predictors: transmissionManual, engine, max_power, age

# 95% Confidence Intervals
```{r}
confint(glm.multiple_final, parm=c("transmissionManual", "engine", "max_power", "age"), level = 0.95)
```

# Test model significance using likelihood ratio test
```{r}
#significance of full model/subset of predictors

glm.null <- glm(selling_price_binary ~ 1,
                data=cars,
                family="binomial")

# Test for significance of the full model:
anova(glm.null, glm.multiple_final, test = "LRT")
```
The p-val of Full model is almost 0. This means that there is there is an improvement on Full model over Null. Our model is significant.


# Will I afford a car with the features I want?
```{r}
#Features I look for in a car

km_driven = 40000
fuel = "Petrol"
seller_type = "Individual"
transmission = "Automatic"
owner = "Second Owner"
engine = 1500
max_power = 120
age = 5

predict(glm.multiple_final, newdata=data.frame(km_driven=km_driven, fuel=fuel, seller_type=seller_type, transmission=transmission,
                                               owner=owner, engine=engine, max_power=max_power, age=age))

predict(glm.multiple_final, 
        newdata=data.frame(km_driven=km_driven, fuel=fuel, seller_type=seller_type, transmission=transmission,
                           owner=owner, engine=engine, max_power=max_power, age=age), 
        type='response')
```
